+++ frontend/src/jsMain/kotlin/ru/posidata/views/utils/externals/telegram/TUser.kt
@@ -5,11 +5,11 @@ package ru.posidata.views.utils.externals.telegram
 
 @JsName("TUser")
 external interface TUser {
-    var auth_date: Int
+    var auth_date: Long
     var first_name: String
     var last_name: String
     var hash: String
-    var id: Int
+    var id: Long
     var photo_url: String
     var username: String
 }
